requirements.txt
groq
python-dotenv
python-telegram-bot